Minimum measure date as 2016-01-01
This was derived from existing production data; there are only 4 K10p
records with dates prior to this; three are possible ATAPS imports (but
not flagged as such in tags), the other looks to be a data error (some
time in 1986).

No k5 or sdq records are before 2016-01-01. Hence starting here as a
sane lower limit on measure date.